requirements.txt
mediapipe==0.9.1.0
mysql_connector_repackaged==0.3.1
opencv_python==4.7.0.72
pygame==2.1.3
pymongo==4.3.3
yacs==0.1.8
